first pass at flattening internal modules
