+++ empty.js
@@ -1,0 +1,6 @@
+//a stream that ends immediately.
+module.exports = function empty () {
+  return function (abort, cb) {
+    cb(true)
+  }
+}
