+++ onEnd.js
@@ -1,0 +1,7 @@
+'use strict'
+
+var drain = require('./drain')
+
+module.exports = function onEnd (done) {
+  return drain(null, done)
+}
