+++ util/id.js
@@ -1,0 +1,1 @@
+module.exports = function id (item) { return item }
