+++ util/keys.js
@@ -1,0 +1,3 @@
+module.exports = function keys (object) {
+  return values(Object.keys(object))
+}
