+++ util/tester.js
@@ -1,0 +1,9 @@
+var prop = require('./prop')
+
+module.exports = function tester (test) {
+  return (
+    'object' === typeof test && 'function' === typeof test.test //regexp
+    ? function (data) { return test.test(data) }
+    : prop (test) || id
+  )
+}
